+++ api/infrastructure/http/model/__init__.py
@@ -1,14 +1,7 @@
 from ._albertmodelhttpclient import AlbertModelHttpClient
+from ._exchanges import FormattedModelRequest, FormattedModelResponse, ModelHttpExchange, OriginalModelRequest, OriginalModelResponse
 from ._mistralmodelhttpclient import MistralModelHttpClient
-from ._modelhttpclient import (
-    FormattedModelRequest,
-    FormattedModelResponse,
-    ModelHttpClient,
-    ModelHttpClientEndpoints,
-    ModelHttpExchange,
-    OriginalModelRequest,
-    OriginalModelResponse,
-)
+from ._modelhttpclient import ModelHttpClient, ModelHttpClientEndpoints
 from ._openaimodelhttpclient import OpenaiModelHttpClient
 from ._teimodelhttpclient import TeiModelHttpClient
 from ._vllmmodelhttpclient import VllmModelHttpClient
